+++ changelog.md
@@ -1,8 +1,9 @@
 ## 1.7.0
 ### Added
-- Charade...
+- Charade words verification with dictionary
 
 ### Changed
+- Charade chances count
 - More tab switches color to be more consistent
 
 ## 1.6.1